Resolve streaming conflicts
