+++ chatapi/src/index.ts
@@ -30,19 +30,17 @@ wss.on('connection', (ws) => {
       data = JSON.parse(data.toString());
 
       if (data && typeof data === 'object') {
-        console.log(data);
+        const chatResponse = await chat(data, true, (response) => {
+          ws.send(JSON.stringify({ 'type': 'partial', response }));
+        });
 
-        // const chatResponse = await chat(data, , true, (response) => {
-        //   ws.send(JSON.stringify({ 'type': 'partial', response }));
-        // });
-
-        // if (chatResponse) {
-        //   ws.send(JSON.stringify({
-        //     'type': 'final',
-        //     'completionText': chatResponse.completionText,
-        //     'couchDBResponse': chatResponse.couchSaveResponse
-        //   }));
-        // }
+        if (chatResponse) {
+          ws.send(JSON.stringify({
+            'type': 'final',
+            'completionText': chatResponse.completionText,
+            'couchDBResponse': chatResponse.couchSaveResponse
+          }));
+        }
       } else {
         ws.send('Error processing input data!');
       }
@@ -54,20 +52,20 @@ wss.on('connection', (ws) => {
 
 app.post('/', async (req: any, res: any) => {
   try {
-    const { data, save, aiProvider } = req.body;
+    const { data, save } = req.body;
 
     if (typeof data !== 'object' || Array.isArray(data) || Object.keys(data).length === 0) {
       res.status(400).json({ 'error': 'Bad Request', 'message': 'The "data" field must be a non-empty object' });
     }
 
     if (!save) {
-      const response = await chatNoSave(data.content, aiProvider, false);
+      const response = await chatNoSave(data.content, data.aiProvider, false);
       res.status(200).json({
         'status': 'Success',
         'chat': response
       });
     } else if (save && data && typeof data === 'object') {
-      const response = await chat(data, aiProvider, false);
+      const response = await chat(data, false);
       res.status(201).json({
         'status': 'Success',
         'chat': response?.completionText,
